pyproject.toml
[project]
name = "chatgpu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents[all]>=1.9.1",
    "strands-agents-builder>=0.1.10",
    "strands-agents-tools>=0.2.8",
    "dbgpu[all]>=2025.8",
]
